Cothread: Build extension with extra flags
This uncovers a (harmless) signed/unsigned mismatch, but is worth doing.

git-svn-id: svn+ssh://serv0002.cs.diamond.ac.uk/home/subversion/repos/controls/diamond/trunk/python/cothread@71667 e099a375-04f9-0310-9d5f-a741eaff62e1

+++ context/cocore.c
@@ -131,7 +131,7 @@ static void save_frame(struct cocore *target)
         frame_size = 0;
     else
     {
-        if (frame_size > target->max_saved_length)
+        if ((size_t) frame_size > target->max_saved_length)
         {
             /* Ensure we have enough room for the required save area. */
             free(target->saved_frame);
